chore: move logger to packages
