+++ app/package.json
@@ -34,7 +34,8 @@
     "react-cookie": "^8.0.1",
     "react-dom": "^19.0.0",
     "react-i18next": "^15.5.1",
-    "tailwind-merge": "^3.2.0"
+    "tailwind-merge": "^3.2.0",
+    "@playlistwizard/logger": "workspace:*"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4.1.4",
