+++ app/src/lib/logger/server.ts
@@ -1,9 +1,9 @@
 import { getEnv } from "@/helpers/getEnv";
-import { type LogLevel, Logger } from "./logger";
+import { type LogLevel, Logger } from "@playlistwizard/logger";
 
 const logLevel = getEnv(["LOG_LEVEL"]).unwrapOr(["info"])[0] as LogLevel;
 
 export const logger = new Logger({
-    name: "Server",
-    level: logLevel,
+	name: "Server",
+	level: logLevel,
 });
